tests/wpt/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-empty-content-value.html

<!doctype html>
<title>CSS Color Adjust Test: meta color-scheme - empty content value</title>
<meta name="color-scheme" content="">
<link rel="help" href="https://drafts.csswg.org/css-color-adjust-1/#color-scheme-meta">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<script>
  assert_root_color_scheme("normal", "Meta color-scheme with empty content attribute has no effect.");
</script>
