third_party/blink/web_tests/external/wpt/css/css-color-adjust/meta/meta-color-scheme-empty-content-value-expected.txt

This is a testharness.js-based test.
FAIL Meta color-scheme with empty content attribute has no effect. assert_equals: expected (string) "normal" but got (undefined) undefined
Harness: the test ran to completion.
